refactor(gh): extract gh/forks submodule
Third of the 5-PR series splitting src/gh/mod.rs into per-resource
submodules (issue #301). Moves the fork-related cluster into
gh/forks.rs: fork counts (parse_gist_fork_counts, gist_forks_plan,
fetch_gist_fork_count(_with)), GraphQL fork-flag detection and
pagination (gist_fork_flags_graphql_query/_plan, the GraphQL response
structs, parse_fork_flags_page, fetch_forked_gist_ids_graphql(_with),
parse_forked_gist_ids_graphql), and fork_of resolution
(gist_detail_plan, fetch_gist_fork_of_id(_with),
collect_owned_fork_of_ids(_with), apply_fork_of_ids,
collect_gist_fork_counts(_with)), along with their unit tests.

parse_gh_gists, GhGist/GhGistForkOf, and escape_graphql_string stay in
gh/mod.rs as shared kernel (also used by comments/gists and stars
respectively); forks.rs reaches them via
`use super::{escape_graphql_string, parse_gh_gists, GhGist};`, matching
the convention established by gh/stars.rs and gh/comments.rs. gh/mod.rs
re-exports forks' public items by explicit name, so none of the
existing crate::gh::X call sites need to change.

No signature or behavior change.

Part of #301
